refactor(front-api): move trigger endpoints out of agent_configurations
Triggers are workspace resources, not a sub-resource of the agent that
happens to reference them. They now live at /api/w/:wId/triggers next to
bulk-execution-mode, with aId as a query param on the collection.

The old paths 307 to the new ones so tabs on the previous bundle keep
working; the redirect goes away in a follow-up.
