+++ front-api/routes/w/[wId]/triggers/[tId]/index.ts
@@ -4,7 +4,7 @@ import executionMode from "./execution_mode";
 import status from "./status";
 import webhookRequests from "./webhook_requests";
 
-// Mounted under /api/w/:wId/assistant/agent_configurations/:aId/triggers/:tId.
+// Mounted under /api/w/:wId/triggers/:tId.
 const app = workspaceApp();
 
 app.route("/execution_mode", executionMode);
